third_party/blink/web_tests/external/wpt/fetch/private-network-access/mixed-content-fetch.tentative.https.window-expected.txt

This is a testharness.js-based test.
Found 52 tests; 41 PASS, 11 FAIL, 0 TIMEOUT, 0 NOTRUN.
[PASS] https-local to http-local: wrong targetAddressSpace "unknown".
[PASS] https-local to http-local: wrong targetAddressSpace "private".
[PASS] https-local to http-local: wrong targetAddressSpace "public".
[PASS] https-local to http-local: not a private network request.
[PASS] https-local to http-private: wrong targetAddressSpace "unknown".
[PASS] https-local to http-private: wrong targetAddressSpace "local".
[PASS] https-local to http-private: wrong targetAddressSpace "public".
[PASS] https-local to http-private: not a private network request.
[PASS] https-local to http-public: wrong targetAddressSpace "unknown".
[PASS] https-local to http-public: wrong targetAddressSpace "local".
[PASS] https-local to http-public: wrong targetAddressSpace "private".
[PASS] https-local to http-public: not a private network request.
[PASS] https-private to http-local: missing targetAddressSpace.
[PASS] https-private to http-local: wrong targetAddressSpace "unknown".
[PASS] https-private to http-local: wrong targetAddressSpace "private".
[PASS] https-private to http-local: wrong targetAddressSpace "public".
[PASS] https-private to http-local: failed preflight.
[FAIL] https-private to http-local: success.
  assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
[FAIL] https-private to http-local: PUT success.
  assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
[FAIL] https-private to http-local: no-cors success.
  assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
[PASS] https-private to http-private: wrong targetAddressSpace "unknown".
[PASS] https-private to http-private: wrong targetAddressSpace "local".
[PASS] https-private to http-private: wrong targetAddressSpace "public".
[PASS] https-private to http-private: not a private network request.
[PASS] https-private to http-public: wrong targetAddressSpace "unknown".
[PASS] https-private to http-public: wrong targetAddressSpace "local".
[PASS] https-private to http-public: wrong targetAddressSpace "private".
[PASS] https-private to http-public: not a private network request.
[PASS] https-public to http-local: missing targetAddressSpace.
[PASS] https-public to http-local: wrong targetAddressSpace "unknown".
[PASS] https-public to http-local: wrong targetAddressSpace "private".
[PASS] https-public to http-local: wrong targetAddressSpace "public".
[PASS] https-public to http-local: failed preflight.
[FAIL] https-public to http-local: success.
  assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
[FAIL] https-public to http-local: PUT success.
  assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
[FAIL] https-public to http-local: no-cors success.
  assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
[PASS] https-public to http-private: missing targetAddressSpace.
[PASS] https-public to http-private: wrong targetAddressSpace "unknown".
[PASS] https-public to http-private: wrong targetAddressSpace "local".
[PASS] https-public to http-private: wrong targetAddressSpace "public".
[PASS] https-public to http-private: failed preflight.
[FAIL] https-public to http-private: success.
  assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
[FAIL] https-public to http-private: PUT success.
  assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
[FAIL] https-public to http-private: no-cors success.
  assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
[PASS] https-public to http-public: wrong targetAddressSpace "unknown".
[PASS] https-public to http-public: wrong targetAddressSpace "local".
[PASS] https-public to http-public: wrong targetAddressSpace "private".
[PASS] https-public to http-public: not a private network request.
[PASS] https-treat-as-public to http-local: wrong targetAddressSpace "private".
[FAIL] https-treat-as-public to http-local: success.
  assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
[PASS] https-treat-as-public to http-private: wrong targetAddressSpace "local".
[FAIL] https-treat-as-public to http-private: success.
  assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
Harness: the test ran to completion.
